refactor(eslint): nest named contract rules
