+++ eslint-local-rules/named-contracts/di/rule.test.ts
@@ -1,5 +1,5 @@
 import { enforceDiFactoryContractsRule } from './rule';
-import { namedContractsFilename, namedContractsRuleTester } from '../named-contracts/testUtils';
+import { namedContractsFilename, namedContractsRuleTester } from '../testUtils';
 
 namedContractsRuleTester.run('enforce-di-factory-contracts', enforceDiFactoryContractsRule, {
     valid: [
